Changed to ES6
