+++ src/modules/localStorage.js
@@ -1,0 +1,4 @@
+export default () => {
+  const listbooks = localStorage.getItem('books') ? JSON.parse(localStorage.getItem('books')) : [];
+  return listbooks;
+};
